feat: added external system component into a separated file
